requirements.txt
llama-cpp-python==0.2.6
requests==2.31.0
yfinance==0.2.36
beautifulsoup4==4.12.3
newspaper3k==0.2.8
python-dotenv==1.0.1
fastapi==0.109.1
uvicorn==0.27.1
pydantic==2.6.1
tenacity==8.2.3
aiohttp==3.9.3 